feat: create car service

+++ src/api/services/cars/DeleteCarsService.ts
@@ -1,10 +1,17 @@
 import { AppDataSource } from 'src/database/data-source-cli';
 import CarsModel from 'src/api/models/CarsModel';
+import { IDeleteCarsEReservesServices } from 'src/api/interfaces/IDeleteCarsEReservesServices';
 
 class DeleteCarsService {
-    public async execute(): Promise<void> {
+    public async execute({ id }: IDeleteCarsEReservesServices): Promise<void> {
         const carsRepository = AppDataSource.getRepository(CarsModel);
-        const cars = await carsRepository.find();
+        const cars = await carsRepository.findOne({
+            where: { id },
+        });
+
+        if (!cars) {
+            throw new Error('cars not found');
+        }
 
         await carsRepository.remove(cars);
     }
